refactor: LikeableEntityType의 위치를 entity에서 enums로 이동
